backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
neo4j
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
supabase
